requirements.txt
Django==2.0
django-filter==1.1.0
django-webpack-loader==0.5.0
djangorestframework==3.7.7
olefile==0.44
Pillow==4.3.0
pytz==2017.3
